move loglevel to library, pass log function to Project on Project Generate
