+++ QgisModelBaker/gui/ili2db_options.py
@@ -23,9 +23,10 @@ from qgis.PyQt.QtCore import QSettings, Qt
 from qgis.PyQt.QtGui import QValidator
 from qgis.PyQt.QtWidgets import QDialog, QSizePolicy
 
+from QgisModelBaker.libs.modelbaker.utils.globals import LogLevel
 from QgisModelBaker.libs.modelbaker.utils.qt_utils import make_file_selector
 from QgisModelBaker.utils import gui_utils
-from QgisModelBaker.utils.gui_utils import LogLevel, get_text_color
+from QgisModelBaker.utils.gui_utils import get_text_color
 
 DIALOG_UI = gui_utils.get_ui_class("ili2db_options.ui")
