+++ QgisModelBaker/gui/workflow_wizard/project_creation_page.py
@@ -586,6 +586,7 @@ class ProjectCreationPage(QWizardPage, PAGE_UI):
             auto_transaction=transaction_mode,
             context={"catalogue_datasetname": CATALOGUE_DATASETNAME},
             optimize_strategy=optimize_strategy,
+            log_function=self.workflow_wizard.log_panel.print_info,
         )
         project.layers = available_layers
         project.relations = relations
